+++ util/options.cc
@@ -811,6 +811,7 @@ ReadOptions::ReadOptions()
       total_order_seek(false),
       prefix_same_as_start(false),
       pin_data(false),
+      background_purge_on_iterator_cleanup(false),
       readahead_size(0) {
   XFUNC_TEST("", "managed_options", managed_options, xf_manage_options,
              reinterpret_cast<ReadOptions*>(this));
@@ -827,6 +828,7 @@ ReadOptions::ReadOptions(bool cksum, bool cache)
       total_order_seek(false),
       prefix_same_as_start(false),
       pin_data(false),
+      background_purge_on_iterator_cleanup(false),
       readahead_size(0) {
   XFUNC_TEST("", "managed_options", managed_options, xf_manage_options,
              reinterpret_cast<ReadOptions*>(this));